Use hips2 in sdt analysis
Summary:
## What

- No change to intraprocedural constraints, though there are some sort order and whitespace differences in the .exp files for `dump` mode. Also, when there are multiple positions that cause something to be NeedsSDT, we keep them all rather than just the first one.
- `dump` mode dumps both intraprocedural and interprocedural constraints. The interprocedural constraint is a placeholder, replaced with more interesting constraints later in the stack.

## Why

This is mostly plumbing to hook up hips2 and sdt_analysis

## Design choices

**Make new version of hips:**

I made a hips-inspired hips2 to:
- reduce coupling between the inter and interprocedural analysis to enable easier iteration on each
- design data structures for scale from the beginning: I'm plan to run it on a large repo very soon

The downside is that if/when we revisit shape-like-dict we would have to tweak it a lot and probably also forward-port features to hips2. There isn't much loss, though, since shape-like-dict didn't scale well to a large repo.

**Use a Tast_visitor.handler instead of a `Tast_visitor.reduce`**

We will need a Tast_visitor.handler our tast visitor, where we will log constraints when running against an entire codebase rather than a single file. The handler format also fit better with the hips2 API, which is why I made the change in this diff.

## What's next

This stack:
- Add a solve mode which copies constraints for each class to their respective descendent classes

After:
- Write constraints from Hh_logger to disk, then do whole-repo solving

> This diff got big, but morally it'sa noop. I couldn't see another way to break the diff up and still have sdt analysis use hips.

Reviewed By: madgen

Differential Revision: D43444646

fbshipit-source-id: d56fbd610c5879dd413dcddf43afbb1c10ce03be