remove "most_recent_file"
Summary:
If hh_server hasn't yet sent us a HELLO to establish the connection, then we want to display its status somewhere (which we read from the status file on disk; we naturally can't obtain it from the persistent connection).

The original design of clientLsp was to display this status as a diagnostic squiggle in line 1 column 1 of the file the user's currently working on, and we used "most_recent_file" to track an approximation of this: the file which most recently received an LSP request.

But D39211079 (https://github.com/facebook/hhvm/commit/dc27c4f9677de6100cf4682cbef487a0002c3cf2) (Sep 22) observed a pathology where VSCode sends LSP requests even for files that aren't truly the active one, so I changed it to no longer display that status on the "most recent file"; instead it now displays it in the alphabetically first file. The comment in that diff said:
> I think the whole concept of "most_recent_file" is dodgy and should be re-examined... either get rid of it entirely if possible, or tame it to only "last opened+edited" file. But that can be left for future work if and when needed.

This diff therefore gets rid of it entirely; it's not being used.

(I'm doing this to streamline the "Lost_server" state, which I intend to use for ide_standalone).

Reviewed By: nt591

Differential Revision: D43590346

fbshipit-source-id: 3a5989106dbac7a2c70fed0c8229e3d69cf3debf