typecheck includes access
Summary:
Last diff banned access from one module to another if they're not in the same package; this diff refines that check by allowing access from one module to another if the package of the former includes that of the latter.

To that end, also refactor `packageConfig` into `packageConfig` and `packages`, where `packageConfig` is solely responsible for loading and parsing the config file on server startup, and `packages` provides general package related APIs.

Reviewed By: jamesjwu

Differential Revision: D43336947

fbshipit-source-id: d0cc57840a8ad07d3417a06af3ae1fa175391165