Remove get_shallow methods
Summary: These were added under the assumption that in some modes, we don't want to provide the typechecker with access to a shallow_decl_provider. But this restriction is historical and shouldn't be necessary after switching over to the populate_member_heaps=false design. Just access these methods via the shallow decl provider instead of requiring implementors of the FoldedDeclProvider trait to write forwarding methods.

Reviewed By: edwinsmith

Differential Revision: D43525547

fbshipit-source-id: cb85865a219a7a45b873dad27e5b73f453df38d7