Move the IDL section to the top level
Summary:
Interface Definition Language was migrated from the wiki and added to Specification because there was a placeholder for the IDL spec. However this section has never been converted into the specification form and we are moving away from specs anyway. Move it to the top level where it makes much more sense and update the links with the following command:

```
grep -rl 'spec/idl' . | xargs sed -i 's|spec/idl|idl|g'
```

Reviewed By: prasad223

Differential Revision: D43586523

fbshipit-source-id: 5cf31233310b2dbb678b6587f5bede4906364bb7