port error_format raw
Summary:
Port OCaml raw_error_formatter.ml to Rust:
* matching color & style
* is_term argument instead of hardcoding stdout
* absolute paths via RelativePathCtx

At least with hh_single_type_check, the path juggling logic is weird enough
that I didn't port it verbatim. In that context, a Pos.absolute's filename is
actually the non-canonical raw path given on the CLI, and not even root
relative. The error formatter attempts to strip `cwd` from the front of this path,
but only works in fragile circumstances.

Reviewed By: nt591

Differential Revision: D43569645

fbshipit-source-id: b4e252a79b3defb1df58dc47db632bd3f89b0cbb