update staticdocs plugin
Summary: In the [post](https://fb.workplace.com/groups/3183360078598952/posts/3516442705290686) the inpage editor doesn't work for one of the pages. When I copy the same text and place it in the editor on the Static Docs site the live preview works as expected. Updating static docs plugin to see if it is the thing to blame.

Differential Revision: D43656126

fbshipit-source-id: 3a6abed205c4b2c3b61c6f00c71bd887eb3c66aa