Return Option from elaborate_raw_id
Summary: When `elaborate_raw_id` returns `Cow::Borrowed`, it currently always returns the input `str`, but it is not constrained to: it could return some substring or a `&'static str`. This is a problem for a caller like `elaborate_id`, which would like to mutate an `Id` only if necessary: we rely on the invariant that the borrowed case always means no mutation is necessary. Communicate this invariant in the type system by returning `Option<String>` instead of `Cow<str>`.

Reviewed By: enetsee

Differential Revision: D43581024

fbshipit-source-id: 684b25a40dc6a489e04801724474c3e283024285