Take &mut Id in elaborate_id
Summary: `elaborate_id` performs an unnecessary allocation in the case that `elaborate_raw_id` returns `Cow::Borrowed` (meaning the given identifier has already been elaborated and does not need to be changed). Accept `&mut Id` so that we can avoid the unnecessary clone of the position (not such a big deal; just a refcount bump) and identifier (which is an unnecessary String::clone in the Borrowed case).

Reviewed By: enetsee

Differential Revision: D43558575

fbshipit-source-id: 794e95bd15d991029ead07ab1e9b0493a3a4d3be