Pull FunDef handling down to visit_fun_def
Summary: This was being done in `visit_def` rather than `visit_fun_def`. That's fine as long as we always enter the visitor via `visit_def` or `visit_program` (as is currently the case), but if we enter the visitor via `visit_fun_def` (as D43512079 and the OCaml visitor do), we fail to set the namespace and type parameters properly. Pull this handling down into `visit_fun_def` (as the OCaml visitor does) so that function declarations can be elaborated individually.

Reviewed By: shayne-fletcher

Differential Revision: D43665057

fbshipit-source-id: f5c8141fb0d3dfd348c7e1767b99bc8381de9a36