Use Function instead of FunctionRef in OnDestructionCallback to make things work on MSVC.
Summary:
The first argument of `EventBase::OnDestructionCallback::schedule` is a
`FunctionRef<void(OnDestructionCallback&)>`.

In `EventBase::runOnDestruction`, we invoke `schedule` with a lambda. On
MSVC, we then get an infinite ping-pong between the `FunctionRef` constructor
in Function.h:1095 and the `isEmptyFunction` call in Function.h:1103.

This looks to be a MSVC-specific issue, because it works fine on other platforms/compilers, but it would still be helpful to fix it.

Reviewed By: Orvid

Differential Revision: D43621191

fbshipit-source-id: 7a73ce1e9899ab584b790b5b2c92a2de99cf6f63