Relax validity checks for refinements
Summary:
When a class constant C is refined on a type X,
we used to check that the bounds given are in
subtype relation with the bounds known on X::C.

This check is not required for soundness and
proved to be annoying to some early users.
Additionally, the check had the side effect of
making the following two types have different
behaviors:

```
X with {type T as T1} with {type T as T2}
X with {type T as T1 as T2}
```

The former could trigger an error (if T2 isn't
a subtype of T1) when the latter would not.

This diff relaxes wellformedness checks. We now
only verify that we refine an existing constant
and that the constant kind (type/ctx) matches.

En passant, I unified the error handling for
these wellformedness checks.

Reviewed By: losvald

Differential Revision: D43619235

fbshipit-source-id: e8ac40488375f95f3f5046bd969a35e412ec9cc2