Distinguish `FinallyBlock`s
Summary:
At least one of our NAST checks requires us to distinguish between a `Block` and the specific case of a block in the finally clause of a `Try` statement.

To translate such passes into our `transform` approach, we need to make this distinction at the type level.

This diff adds a `FinallyBlock` newtype and updates the lowerer and emitter correspondingly

Reviewed By: shayne-fletcher

Differential Revision: D43658903

fbshipit-source-id: f37e7c5074bde383618601939f5f9e2a699c1147