ide_standalone shouldn't connect to hh_server
Summary:
An earlier diff added a JK for ide_standalone, `Serverless_with_shell`, which means HackIDE will no longer use a persistent connection to hh_server.

In this diff I respect that flag by
1. not attempting to form a persistent connection to hh_server
2. having the status-bar indicator solely display HackIDE status

I chose to represent `Serverless_with_shell` with the clientLsp state Lost_server. I envisage, once we've completed the migration to Serverless_with_shell, to have these states:
```
  | Pre_init  (** Pre_init: we haven't yet received the initialize request. *)
  | Lost_server of Lost_env.t
      (** Lost_server: this is used for (1) Serverless_with_shell which means
          that we never want an hh_server connection, in which case
          Lost_env.params.new_hh_server_state will always be None,
          or (2) Other modes which do want a persistent connection, but either
          we failed to start hh_server up, or someone stole the persistent connection
          from us, in which case we might choose to grab it back if prompted... *)
  | Post_shutdown
      (** Post_shutdown: we received a shutdown request from the client, and
          therefore shut down our connection to the server. We can't handle
          any more requests from the client and will close as soon as it
          notifies us that we can exit. *)
```
In other words, it will no longer have the two states specific to hh_server connection status:
```
  | In_init of In_init_env.t
      (** In_init: we did respond to the initialize request, and now we're
          waiting for a "Hello" from the server. When that comes we'll
          request a permanent connection from the server, and process the
          file_changes backlog, and switch to Main_loop. *)
  | Main_loop of Main_env.t
      (** Main_loop: we have a working connection to both server and client. *)
```

I decided to use "Lost_server" because it's already how clientLsp indicates that it's running normally but without an RPC connection to hh_server.

Lost_env.t has the following field of interest:
```
    new_hh_server_state: hh_server_state;
```
In this diff, I decided to make that field an option, and use None to indicate that the reason for Lost_server state is that we never even wanted to connect to the server in the first place. This "None" value is what the diff uses to drive the way it merges status -- it will merge status from clientIde with "None" status from hh_server.

I expect that clientLsp should support only a single shell-out-to-hh-server at a time, and Lost_env.t will eventually include a promise/callback indicating what the current shell-out is at the moment. In this way, in future, clientLsp will be able to remain active even while an hh_server shell-out is underway.

Reviewed By: nt591

Differential Revision: D43591813

fbshipit-source-id: 762e080efb7fe1b6d64ccd560f392766d48d84de