Delay loading repo options when `LoadFilepathFromUnitCache` is set
Summary: `--hhas-round-trip` tests compile a Hack file, emit it as HHAS, then assemble it from disk to run it. `EvalLoadFilepathFromUnitCache` allows us to emulate the original file path, but this was not being respected for loading repo options. Hence, when attempting to turn on NF via `.hhvmconfig.hdf`, we'd fall over.

Reviewed By: edwinsmith

Differential Revision: D43631090

fbshipit-source-id: 6e91aa85504c18c27e004cda42478444a5271de2