supportdyn pushing for Traversable
Summary:
Add supportdyn pushing for KeyedTraversable and for Traversable. That is `supportdyn<Traversable<t>> <: Traversable<supportdyn<t>>`. This is sound because the only way to get a value out of a Traversable is to call no-argument methods, e.g., `current() : T` from Iterator, and if the Traversable was supportdyn, then any method in it must return SDT values when all 0 of the the arguments are SDT.

We don't currently support supportdyn pushing for non-Container subclassese of Traversable/KeyedTraversable. We could add support for some builtins (e.g., Iterator) on a case-by-case basis

Reviewed By: andrewjkennedy

Differential Revision: D43696984

fbshipit-source-id: 44fe36cb709da48af84b7d037a53c794bee23a34