Use variation on ordinary subtyping for override checks
Summary: Under legacy dynamic, we are more strict in the subtyping used for override checks that used for parameter passing and return. In particular, we do not allow a method declared `foo():int` to override one declared `foo():dynamic`. Under Sound Dynamic, it's safe to permit this, but we do not want to use the full "dynamic-aware" subtyping relation, as this would permit `foo():~int` to override `foo():~string`. So instead, we do the following: we allow a return type of `t` to override a return type of `u` *either* if `t <: u` *or* `u` is exactly type `dynamic`, and `t <:D dynamic` (i.e. a dynamic aware subtype of dynamic). We do the same thing for parameters, but flipped around because they behave contravariantly. For returns we also allow `Awaitable<t>` to override `Awaitable<dynamic>` or `~Awaitable<dynamic>` if `t <:D dynamic`.

Differential Revision: D43504896

fbshipit-source-id: 219ba8cf3f59a7aecffb8ac5643320d07548e44e