HH\ImplicitContext\get_state_unsafe
Summary:
In theory, observing which IC state you are in breaks memoization cache poisoning guarantees that the memoization options are meant to enforce. In practice, sometimes you want to bypass the restrictions in a controlled manner e.g. to effect custom logging or simulate behavior in user land that would normally be built-in.

This diffs adds a function for this purpose which tells you which IC state is currently active (but not the actual value in the value state).

One example of use is if you want to remove a backdoor from code that may be called from a variety of states, you would need to replace the backdoor with a call that converts calls from a value state to the soft_set state and calls from an inaccessible state to a soft_inaccessible state.

Reviewed By: rodmk

Differential Revision: D42488970

fbshipit-source-id: a14c71fb1fc2a9cd35d423300fe5c15405dfaf7e