Add Field adapter structs
Summary:
This adds structs to testset.thrift for field adapters, as well as corresponding adapters for python, c++, to wrap thrift fields.

All of the adapters are noops that don't really use field adapter features because it'd be too complicated otherwise.

Also noting that I changed the resulting type in the java tests from Wrapper to Wrapped because Wrapper is the java term for field adapter and it felt annoying to have to deal with 2 clashing terms.

Reviewed By: Mizuchi, bahadird

Differential Revision: D42724236

fbshipit-source-id: 644f692b1a61cbd38cba4466db7213679cd6d3f0