Check if final constructors are overidden in `__ConsistentConstruct` hierarchies
Summary:
Today, this code passes Hack and fatals unconditionally on load in HHVM:
```
<?hh

<<__ConsistentConstruct>>
abstract class Foo {
  final protected function __construct() {}

}

final class Bar extends Foo {
  protected function __construct() {
    parent::__construct();
  }
}

```

A comment in `typing_extends.ml` implies that we check this elsewhere. That *might* have been true at one point. Before this diff the logic for `__ConsistentConstruct` was:
```
if class is `__ConsistentConstruct`
  run all checks
else
  check if overriding a final constructor
  check if overriding with an invalid visibility
```
`run all checks` *skips* the final override check, but not the visibility check, for non-methods. I think in the future we can simplify this further, but I'd rather get this work in first.

Reviewed By: vassilmladenov

Differential Revision: D42114119

fbshipit-source-id: ced329dfd0a0c702bd3f483082e01936e15c18d5