add config to logger struct
Summary:
Add `Config` to `FileScubaLogger`, so loggers always store their associated config. This is an API change rather than a functionality change -- logging should behave similarly to before. This should make it easier to pass a logger around without having to pass or create its associated config when writing a config file.

Most of the changes are in `file_scuba_logger.rs`. Changes include making `file` and `scuba` private, which is responsible for most of the changes in other files.

Reviewed By: ljw1004

Differential Revision: D43096307

fbshipit-source-id: 471d7387eb25e2accfadf90dcbb3b4a1a3a2bdcc