switch test logger to stderr
Summary:
The logger we've been using for buck tests so far print logs to stdout. This becomes a problem if we want to expose logs from the worker binary. This diff switches that over to stderr.

For why we need to do this, see the comment in the implementation of `write()` for `TestStderrWriter` in `hh_slog.rs`.

Reviewed By: edwinsmith

Differential Revision: D43152763

fbshipit-source-id: ac524ecd2cb0de80e5a7c26bc4cbfe271c0ba0c3