Allow ChannelProcessor to re-use existing sequenced executors
Summary:
ChannelProcessor currently creates its own set of SequencedExecutors around the user-provided executor. This can use a lot of memory if an application has many ChannelProcessor instances.

This diff allows the user to optionally provide a pool of sequenced executors to use, allowing multiple ChannelProcessors to share the same pool.

Reviewed By: aary

Differential Revision: D43342417

fbshipit-source-id: 79acae53668cd5569fdf1c7a0ea2e3521b25e2cb