FanoutSender optimizations
Summary:
Previously, rather than listen for callbacks indicating that subscribers have cancelled (and deleting the corresponding senders from its set), FanoutSender would iterate over all senders and check to see if the corresponding receiver still existed. This was done to save memory. However, it causes subscribe operations to be O(# senders), rather than O(1).

This change makes FanoutSender listen for cancellation callbacks and delete the corresponding senders from its set, but only in the case where there are 2 or more subscribers. This way, we preserve the property that a FanoutSender is no more expensive than a normal Sender when there is only one subscriber.

Reviewed By: aary

Differential Revision: D43342420

fbshipit-source-id: 687e3319af0914cb20d1867df502e98c75734632