derive show instead of using custom debug function
Summary:
to avoid having to add `debug_` functions for everything in the `Intra` module, derive `show`.

The intent of the `show` is just for debugging.
The downside of the derive is that hips2 might accidentally depend on `show`ing intraprocedural constraints. I vote for choosing conveience and readable API over having a bunch of explicit `debug_` functions, since I think this will save the most time in the long run.

Reviewed By: madgen

Differential Revision: D43696132

fbshipit-source-id: 05fbb52648816812c21cdb13769d1a509292e55c