add infra for custom interprocedural constraints and add constraint `SyntacticallyNadable`
Summary:
Add a new interprocedural constraint. This change only affects `dump` mode in the current diff.

The new constraint is used for summarizing results in the next diff in the stack. Enums and enum classes cannot have the `__NoAutoDynamic` attribute, so we shouldn't report that they are `__NoAutoDynamic`-able.

Since the concept of syntactic `__NoAutoDynamic`-ability is domain-specific, introduce new hips2 API for custom (domain-specific) interprocedural constraints.

Reviewed By: madgen

Differential Revision: D43696134

fbshipit-source-id: 83876d79e17664168cbb33ba38c29b4e700f3c47