defer loading key-value tast handlers
Summary:
# What

Only create the optional tast handlers we need.

# Why

In general, this avoids creating unneeded state and running unneeded code.

The specific reason for doing this now is that I'm planning to add persistence for sdt analysis (later in the stack). We need to know when to flush to disk, and a tast handler doesn't know when it's "done". So we attach an `at_exit` hook to flush to disk. To avoid messing with our main `hh` use cases, it's important that this flushing code only happen when the sdt_analysis tast handler is selected.

Reviewed By: madgen

Differential Revision: D43618795

fbshipit-source-id: 53dd08a350f31d42f985a2b1ca8b1f949db229fa