Fix crash in function coverage logic
Summary:
When collecting function coverage the JIT uses `CheckFuncNeedsCoverage` to decide if it needs to collect coverage. This instruction checks if the `s_coverage_index` value is non-zero (indicating that coverage is enabled) and that it is not equal to the per-function coverage index (indicating the current function has not yet been recorded).

The cleanup logic for function coverage was resetting `tl_called_functions` but neglected to set `s_coverage_index` back to zero.

Differential Revision: D43779818

fbshipit-source-id: bcb5505a8205d2e6ffc79cb809006bcdf13aba82