Use exact code to sort error list before deduplicating
Summary:
We divide the error code by 1000 to get the phase e.g. 2 = naming, 4 = typing, but we use this for both the compare function for sorting and the equality function for deduplicating. This makes the sort function unstable. The unstable sorting function caused hh_single_type_check and hh_server to de-duplicate errors differently.

This makes use compare the exact code for sorting.

Reviewed By: ljw1004, edwinsmith

Differential Revision: D43774369

fbshipit-source-id: 35f2cf98f3b43783914073d0c81351a0c7694202