Use third-party/tbb to build HHVM
We currently don't use the TBB in third-party/tbb
for building HHVM, which is causing some issues with
the Ubuntu 22.04 update.

This PR updates our CMake files to use the TBB in
third-party. This uses `ExternalProject_add` to build
the existing code in a bundled target. This complemented
with an interface target that exports the static library
to the rest of HHVM.

The approach eschews TBB's CMake recommendation (a
custom function called `tbb_build`) as this seemed
poorly integrated and adds more opacity to our build.
Instead we just call `make` directly to build the code.

This removes the need for calling `find_package(tbb)`
so FindTBB.cmake was removed too.