Port `nast_generics_check.ml`
Summary:
Reimplements `nast_generics_check`; the pass checks type parameters for shadowing and illegal names. Under higher-kinded types, it will also check for non-shadowed reuse of names and a range of features non-supported in combination with HKT, and refences to HKTparams in `where` constraints.

The OCaml version of this uses two visitors the second of which is a top-down traversal of hints. This version combines them into a single transform pass so has additional information in the context to determine if we have a hint in a `where` constraint on a function or method.

Reviewed By: shayne-fletcher

Differential Revision: D43735880

fbshipit-source-id: a47dab03dc5509966e619ffe2f89242af8967310