Move `Tparam_non_shadow_reuse` from `Typing_error.Primary.t` to `Naming_error.t`
Summary:
The `Tparam_non_shaow_reuse` error constructor is only ever used during NAST checks.

It seems to have been added as a typing error since it refers to errors arising with higher kinded type but we should be classify by the phase in which the error is generated.

Moving the error will allow us to raise it in Rust-based elaboration

Reviewed By: shayne-fletcher

Differential Revision: D43731583

fbshipit-source-id: 82dfcf9aa69093b8af242758da2a0cb2d0680ce3