Document whitespace and comments
Summary:
Comments and whitespace in Thrift are **not** identical to comments and whitespace in C++, there is no such thing as `##` comments and we don't plan to remove support for any kind of comments right now.

Remove all of these inaccuracies and instead document what whitespace and comments actually are without trying to refer to other languages.

Rename the section to Whitespace and Comments and move to the beginning to follow the convention of having simpler and more common concepts first. It is also customary in other references to discuss this at the beginning, see e.g. https://docs.python.org/3/reference/lexical_analysis.html and https://docs.swift.org/swift-book/documentation/the-swift-programming-language/lexicalstructure.

Also add a few examples of literals.

Reviewed By: iahs

Differential Revision: D43805964

fbshipit-source-id: 54e59ee310914000324abac85e130db12fc3c6b0