Remove assertion in typedef lookups in LazyFoldedDeclProvider
Summary:
The default implementation of `get_class` and `get_typedef` in hackrs' FoldedDeclProvider perform a naming table lookup to determine whether the given type name corresponds to a class or typedef declaration. But this differs from hh_server's behavior: when asked for a class, hh_server will return a value from its classes store, even if the naming table says that class' name corresponds to a typedef. The reverse is also true: when asked for a typedef, if both a typedef and class exist with that name, we'll return the typedef even if the symbol table says the class is the canonical definition.

rust_provider_backend, on the other hand, only returns a class when the naming table claims the given type name corresponds to a class (and likewise for typedefs). When the naming table claims the given type name corresponds to a typedef, but we are unable to get a typedef of that name from the shallow decl provider, we crash.

I don't have a good understanding of the scenarios where this behavior difference arises in practice (perhaps a race condition where the file is edited after the naming table lookup), but we have [telemetry indicating that it does happen](https://fburl.com/scuba/hh_server_events/g8j6zl55), causing crashes when rust_provider_backend is enabled.

I attempted to remove the naming table lookup (in an earlier version of this diff), but found that this causes sandcastle_perf runs to fail via OOM. I'm not sure why this is. For now, we can leave the naming table lookup in, and address the errors by quietly returning None instead of raising a panic. This seems preferable to the hh_server behavior anyway (i.e., we return decl results which are more consistent with the naming table).

Reviewed By: shayne-fletcher

Differential Revision: D43715531

fbshipit-source-id: d3d8833eff7e1e92117b96a8609064759f261d3f