Attach File Attributes to Module Declarations
Summary:
# Context
File Attributes are implemented by attaching the attribute to top level declarations. When the top level declaration is type checked, the attached file attributes will be type checked as well. When File Attributes were added the only top-level declarations that existed were:
* functions
* type defs (`newtype`, `type`)
* class (`class`, `interface`, `trait`)

When modules were introduced to the language this added a new kind of top level declaration. Unfortunately File Attributes were not being attached to modules. This meant that if a file used file attributes and only contained modules it would not get type checked. For instance:

```
<?hh
<<file: NonExistentAttribute>>

new module foo {}
```

Would not produce an error, but
```
<?hh
<<file: NonExistentAttribute>>

class Foo {}
```

would

# Mitigation
I fixed this by adding a file attributes field to the aast for modules and attaching the file attributes during namespace elaboration just like we do for the other top-level declarations.

In the future we should change how File Attributes are checked so they are checked regardless what kind of top-level declarations are present in the file.

Reviewed By: shayne-fletcher

Differential Revision: D43718979

fbshipit-source-id: d501d6a8ef458b1331b07c6254d1fb4b20224115