Add flag for skipping check under dynamic assumptions
Summary:
For functions and methods marked `<<__SupportDynamicType>>` the typechecker must in principle check the body under dynamic assumptions for parameters and return. Often this comes "for free" because parameters are enforced, and the body is checked as normal but with `supportdyn` wrapped around the return type. But if not, we do a second check of the body.

The new flag `--skip-check-under-dynamic` lets you skip this second check. This is to be used only for test, particularly for performance reasons to isolate costs of Sound Dynamic.

Reviewed By: losvald

Differential Revision: D43852514

fbshipit-source-id: 71f08b47ce31b3e74e9bcbc887475af93203dcde