supportdyn refinement
Summary: When doing `$x as T<_>`, if `$x <: supportdyn<t>`, then it is sound for the refinement to intersect on `supportdyn<T<T#1>>`. This opens up the opportunity for the `supportdyn` to be further pushed inside of `T` if it is a `Traversable` or subtype of `Container`. If the type of `$x` is directly `supportdyn<t>` the intersection already does this, but if `$x` is a varaible or generic, it won't, so we pick it up in the refinement code.

Reviewed By: andrewjkennedy

Differential Revision: D43767403

fbshipit-source-id: a775330fe80e9a927a262d0028f8866ad4024ac1