Update function env in TAST with correct assumptions
Summary:
In D43462746, we changed the environment to distinguish if typechecking is done once or twice due to sound dynamic and if so under which assumptions.

When methods are typechecked twice we were correctly updating the environment of the method with normal assumptions to correctly reflect the assumptions. However, the relevant bit of top-level function check wasn't doing that.

This diff makes the double typecheck behaviour for top-level functions match that of methods.

Reviewed By: andrewjkennedy

Differential Revision: D43625394

fbshipit-source-id: e39cde2c273d3ca327002685c12f0cc79033cd72