Translate NewObjD directly as __sil_allocate
Summary:
With NewObjD we know exactly the type of the object to allocate. There's no
need to dispatch to a `__factory` method through singleton object, and instead
we can just use textual's typed allocation instruction.

Alternatively, we could lower `NewObjD` into a direct call to `A$static__factory` but then we'd also need to provide either a definition **or** a declaration because the referenced class can be defined in another translation unit. Going straight to __sil_allocate seemed like a more straightforward path.

Reviewed By: aorenste

Differential Revision: D43731326

fbshipit-source-id: b817e00c5e0b4a678da36d05a70335b0ef0236a5