service: allow debugInodeStatus to not be recursive
Summary:
The debugInodeStatus can be used to get a mapping between paths and objectid
and/or inode numbers. Unfortunately, the Thrift call is recursive by default,
so calling it on the root of the repository yields a potentially extremely long
Thrift call.

By adding a non-recursive flag, we can mitigate this and force EdenFS to merely
return data for a single directory. This will be used in doctor to obtain the
mapping between paths and objectid in an efficient way. In the future, the
Windows-only filesystem checker in doctor will also be migrated to use the
non-recursive API to reduce races during doctor.

Reviewed By: genevievehelsel

Differential Revision: D43859190

fbshipit-source-id: e175238c2a169b5cdfcf9d4dbcb058578d0c2efe