Dead code removal - warnings 8xxx
Summary:
In D8237813 (June 2018), fredemmott introduced "forward_compatibility_level" option in .hhconfig. The idea was that we would alert the user about forward compatibility. The next diff in the stack D8240015 added the distinction between warnings and errors so that forward compatibility could be classed as a warning. (prior to the diffstack, every error in Errors.t had by assumption been an error).

In D14087451 (Feb 2019), fredemmott removed "forward_compatibility_level" option. He wrote: "Unused since October, and makes automated open source releases more complicated."

This diff does the remaining part of removal: it removes the distinction between warnings and errors.

(I noticed this while I was doing unrelated work on LSP. I noticed that the code looked at the error code to determine whether something was a warning or an error, and got suspicious since I'd never even seen a warning, and dug in further).

Reviewed By: Wilfred

Differential Revision: D43883325

fbshipit-source-id: ebf944b9af4dd62c4125c4ee3500d70608142acc