Fix bug with type parameter environment overriding
Summary:
Previously, I was using the wrong API to retrieve the type parameter environment which happened to have the same behaviour as the most conservative version of the check I was trying to implement.

This diff changes it so that the intended behaviour of selective `supportdyn<mixed>` bound removal happens. This in turn prevents us from aggressively marking definitions with `NeedsSDT`.

"before this change, `doesnt_need_sdt` in the test was marked as `NeedsSDT`"

Reviewed By: mheiber

Differential Revision: D43878318

fbshipit-source-id: cb1b31ead5ef1fdee2bdbc3e02536abee7e0a8f0