Formation check based on returns
Summary:
If a definition is marked with `NAD` the explicit returns of callables need to match the non-pessimised type of the callable.

This diff strips the like type of the return type stored in the callable (which will work so long as we are in the pessimised world) and checks it against the type of the expression in return statements.

Reviewed By: mheiber

Differential Revision: D43886667

fbshipit-source-id: 38152137cd4de1c1e625c135c7e3d50d90c80422