Remove unnecessary supportdyn bound from Hack arrays
Summary:
Remove the `as supportdyn<mixed>` bound from all Containers/KeyedContainers that were used for implicit pessimisation.

Refinement now relies on supportdyn pushing, i.e. `supportdyn<vec<mixed>> <: vec<supportdyn<mixed>>`. Consider `$x : supportdyn<mixed>`. Previously, `$x as vec<_>` would leave `$x` as type `supportdyn<vec<T#1>>`, where `T#1 <: supportdyn<mixed>`. The bound on `T#1` came from the bound on `vec`, which is now missing, so we need to push the outer suportdyn to recover that the contents of an SDT vec are SDT.

This leaves the .hhis in a state where there are errors under `--enable-sound-dynamic-type` if they aren't pessimised (either explicitly or implicitly). This is because KeyedContainer &co have `<<__SupportDynamicType>>` annotations in the .hhi, and some subclasses don't, since they'll get them under pessimisation. To also pessimise these subclasses in the hhis will require lots of like type markers so as to not introduce like types when `--enable-sound-dynamic-type` is not on

Reviewed By: andrewjkennedy

Differential Revision: D43312774

fbshipit-source-id: b00992631c7307a21053ac5042273d353aa98068