Move some tests that depend on watchman to static file watcher
Summary: Some autoloading tests depend on watchman insofar as they need a way to ask for changed files. For tests where the entire repo is static that can be replaced with a static file watcher. Let's make that change and move the affected tests into the cmake build as well.

Reviewed By: edwinsmith

Differential Revision: D43882445

fbshipit-source-id: 2f8875304da1fd803bd43b673f1bc73aaf2f25a5