Fully resolve type-constraints using extern-worker
Summary:
Right now when HHBBC needs to resolve a type, it walks through any
type-aliases or enums to fully resolve it. This is somewhat wasteful,
as it may need to walk the alias chain multiple times. After some
analysis, we'll update some type-constraints (but not all) with the
resolved type.

This kind of "on-demand" type resolution is difficult when working
distributed, as all of the types might not be available, and it can be
cumbersome to infer which ones are needed. Instead of doing it
on-demand, pre-resolve all type-constraints early on. Once the
type-constraints are resolved, their contents can be fully trusted. No
other types will need to be walked through.

When parsing the units, we record all type-aliases (enums are already
stored by being virtue of being a class). For every class and every
func, we also record the types needed by all type-constraints.

As a first step, we "flatten" all type-aliases and enums, constructing
a mapping from those to their ultimate resolved type. We might detect
invalid type-aliases or bad enums here too.

During the standard FlattenJob, we also provide any type-aliases or
enums referenced by the input classes (the information gathered during
parsing). In addition, we provide the names of the classes which were
referenced by type-constraints, but *don't* exist. This is more
efficient than providing the list of classes which do exist, so
references to non-existent classes are much rarer. The FlattenJob uses
this provided information to update the type-constraints. If a name
doesn't correspond to a type-alias or an unknown class, it is assumed
to be an object with that name.

After this step, any type-constraint which remains
AnnotType::Unresolved corresponds to a missing type (so
Bottom). Likewise, if it is resolved, it corresponds to a valid,
existing type. However, the FlattenJob might discover that certain
classes are not instantiable during processing. We might have already
rewritten type-constraints to point to that class. To deal with this,
we have a "fixup" pass afterwards. If any classes were discovered to
be uninstantiable, we'll do a pass to fixup the related
type-constraints by making them unresolved. This is rare in practice.

Doing all of this lets us remove the pass to resolve property
type-constraints and also remove some code to resolve type-constraints
in the Index.

Reviewed By: jano

Differential Revision: D43893505

fbshipit-source-id: d3b8fab35e882b6df6e0abe1249e4bc039aeefb5